Try to avoid long type int type conversion, so change type from long to int

+++ Factorial.h
@@ -1,4 +1,4 @@
-long Factorial(long n){
+int Factorial(int n){
   if (n==1){
     return 1;
   }
